requirements.txt
vapi_server_sdk

flask

python-dotenv

pyngrok

uvicorn